Correções menores

+++ src/main/java/com/kuuhaku/command/commands/discord/misc/SellCardCommand.java
@@ -65,10 +65,7 @@ public class SellCardCommand implements Executable {
 		Kawaipon kp = KawaiponDAO.getKawaipon(author.getId());
 		Deck dk = kp.getDeck();
 
-		if (dk.isNovice()) {
-			channel.sendMessage("❌ | Você não pode fazer esta operação com o deck de iniciante!").queue();
-			return;
-		} else if (args.length < 2) {
+		if (args.length < 2) {
 			channel.sendMessage("❌ | Você precisa informar uma carta e o preço dela.").queue();
 			return;
 		} else if (!StringUtils.isNumeric(args[1])) {
@@ -202,7 +199,7 @@ public class SellCardCommand implements Executable {
 								Kawaipon finalKp = KawaiponDAO.getKawaipon(author.getId());
 								Deck fDk = finalKp.getDeck();
 
-								if (fDk.isNovice()) {
+								if (fDk.isNovice() && off.getMiddle() == CardType.SENSHI) {
 									s.delete().flatMap(d -> channel.sendMessage("❌ | Você não pode fazer esta operação com o deck de iniciante!")).queue();
 									Main.getInfo().getConfirmationPending().remove(author.getId());
 									return;
